Added an extra feature of the preview board
The previous version utilises the player one's board to show a preview
of ship placement before game start. The preview board is created to
avoid interface with the actual player board because the preview board
created and updated based on user's input in the DOM level. The actual
player board however is created and rendered based on the board property
stored within the player instance.