Added the logic to transit from the preview board for player input to the actual gameplay
Added a feature to generate a valid but random board  for the computer
player. When the user inputs are submitted and validated, the page will
transit to rendering the real player's board and the computer's randomly
generated board. From there onwards, the game logic begins which the
event is listened through clicks on the computer's board.