Fixed game logic bugs
There was a bug where player can click on a cell where ship has already
been hit. The bug was fixed in the receiveAttack method to change the
array entry to false only if the array entry is null.

There was another bug where player can press on the same cell of the
opponent's board and his board will still receive attack automatically.
The bug was fixed by adding a condition to make sure that the cell is
neither a true or false when its coordinate is translated to the array.