Added drag and drop feature for the starting page
Drag elements are created as a a flex container containing the number of
cells representing the ship size. A event listener is attached to the
drag item to detect which cell has been clicked on and another event
listener is attached to the same drag item to detect the item that is
being dragged and store its data temporarily with the setData() method.

Drop zones are created for cells of the preview board. Event listeners
are attached to each of these drop zone cells to detect which cell the
user will drop the drag item in. Once the drag item is dropped, the
getData() method is used to retrieve the ship data including direction
and ship size in order to determine the starting coordinate of the ship
placement. The starting coordinates will then be transfered to the input
fields which preview board will be updated automatically.